Took at look at graph clustering with IRM.
There are some problems with the interpretation of our case with respect to graph clustering and IRM, but there must be a way around it. I probably just haven't understood Mikkel's paper well enough. I also included the library demos, so Fabian has a chance to learn  how to operate that library.